requirements.txt
fastapi
uvicorn
pandas
python-dotenv
requests
google-generativeai
